gun_control/requirements.txt
Flask
flask-socketio
gevent
numpy
pillow
pyserial
RPi.GPIO
picamera2